requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.21.6
openai==0.27.8
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.23.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
yarl==1.9.2
zipp==3.15.0
